requirements.txt
dnspython==2.5.0
kafka-python==2.0.2
PyYAML==6.0.1
